requirements.txt
debugpy
faster-whisper>=1.2.1
PyYAML
requests
tqdm>=4.66
